+++ packages/core/src/index.ts
@@ -1,8 +1,9 @@
-export { BaseTranslation } from './core'
-
-export { LocaleTranslations, FormattersInitializer, AsyncFormattersInitializer } from './util.loader'
-
-export { i18nString } from './util.string'
-export { i18nObject } from './util.object'
-export { i18nObjectLoader, i18nObjectLoaderAsync } from './util.loader'
+export { BaseTranslation, LocalizedString } from './core'
 export { i18n } from './util.instance'
+export { AsyncFormattersInitializer, FormattersInitializer, i18nObjectLoader, i18nObjectLoaderAsync, LocaleTranslations } from './util.loader'
+export { i18nObject } from './util.object'
+export { i18nString } from './util.string'
+
+
+
+